Addressed several comments, still todo - separate sensor classes.
